requirements/main.txt
aiobotocore == 2.9.0
bcrypt == 4.1.2
cryptography
databases[asyncmy] == 0.8.0
datadog == 0.49.1
ddtrace == 2.7.6
email-validator == 2.0.0
fastapi == 0.108.0
fastapi-limiter == 0.1.5
httpx == 0.26.0
meilisearch-python-sdk == 2.0.1
orjson == 3.9.15
python-dotenv == 1.0.1
python-multipart
redis == 4.6.0
types-aiobotocore[s3] == 2.5.2
uvicorn == 0.19.0
uvloop == 0.19.0; sys_platform != "win32"
winloop == 0.1.0; sys_platform == "win32"
xor-cipher == 3.0.1
python-json-logger == 2.0.7
PyYAML == 6.0.1